requirements/base.txt
Django==2.1.2
django-model-utils=3.1.2
django-environ=0.4.5
djangorestframework=3.8.2
